requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
